improve llm module
