+++ ankiops/llm/__init__.py
@@ -1,6 +1,8 @@
 """OpenAI structured-output LLM support for AnkiOps."""
 
-from .runner import list_jobs, plan_task, run_task, run_task_async, show_job
+from .execution import run_task, run_task_async
+from .jobs import list_jobs, show_job
+from .planning import plan_task
 
 __all__ = [
     "list_jobs",
